moved game Bubbles to the games package
