+++ gazeplay/src/main/java/net/gazeplay/DefaultGamesLocator.java
@@ -2,6 +2,7 @@ package net.gazeplay;
 
 import creampie.CreamPie;
 import lombok.extern.slf4j.Slf4j;
+import net.gazeplay.games.bubbles.BubblesGamesStats;
 import net.gazeplay.games.magiccards.Card;
 import net.gazeplay.games.blocs.Blocs;
 import net.gazeplay.games.blocs.BlocsGamesStats;
