+++ gazeplay/src/main/java/utils/games/stats/StatsDisplay.java
@@ -16,6 +16,7 @@ import javafx.scene.paint.ImagePattern;
 import javafx.scene.shape.Rectangle;
 import javafx.scene.text.Font;
 import javafx.scene.text.Text;
+import net.gazeplay.games.bubbles.BubblesGamesStats;
 import utils.games.HeatMapUtils;
 import utils.games.HomeUtils;
 import utils.games.multilinguism.Multilinguism;
